Add Web QA UI tests to Travis-CI

+++ tests/serve.sh
@@ -4,7 +4,11 @@ pushd /tmp/marketplace-api-mock
 pip install --user --exists-action=w --download-cache=/tmp/pip-cache -r requirements.txt
 python main.py &
 popd
-mv src/media/js/settings_local_test.js src/media/js/settings_local.js
+if [ $FIREPLACE_SETTINGS ]; then
+    mv $FIREPLACE_SETTINGS src/media/js/settings_local.js
+else
+    mv src/media/js/settings_local_test.js src/media/js/settings_local.js
+fi
 make build
 MKT_COMPILED=1 make serve &
 sleep 10
